use SAF by default on API level 19 and up
this changes the default value for the "Storage Access Framework" setting
to true on kitkat and higher.
the legacy way of opening files(ACTION_GET_CONTENT) causes some problems
on modern android versions, and i think it would be a good idea to use
ACTION_OPEN_DOCUMENT if it is available.