MagicKeyboard: Added key to change password
Added a key to ease the process of changing a password, by retrieving the
previous password from an entry's hisotry.

There are many reasons wanting to change a password even when using a password
manager, for example, for expired passwords, known breaches or changing to a
safer password after migrating to KeePassDX.

It's not clear or intuitive how to change change a password when using the
MagicKeyboard. This change introduces a key to retrieve an entry's old
password for a streamlined procedure:
 1. Generate and save a new password for an entry in the keepass DB
 2. Open the change password form of the target website/app
 3. When requested to enter the old password, use the "Previous Password" key
 4. Enter the new password using the normal password key

This process is resilient to errors, if the process timeouts or the password
is rejected, the old password is still stored in the database. If the new
password is rejected, the user can change it on the main app, save it, delete
the last history entry (with the rejected password) and repeat.